convert sealed interface and implementers to single enum
