--- src/main/java/com/xyzcorp/javapatterns/state/functional/Advantage.java
@@ -1,14 +1,0 @@
-package com.xyzcorp.javapatterns.state.functional;
-
-public final class Advantage implements Score {
-    @Override
-    public Score wins(Score opponentsScore) {
-
-        return new Won();
-    }
-
-    @Override
-    public Score loses(Score opponentsScore) {
-        return new Forty();
-    }
-}
